Working commit, move out getInitialFormValues of PhoneEditView
