+++ src/applications/personalization/profile/util/getInitialFormValues.js
@@ -1,0 +1,28 @@
+export const getInitialFormValues = (type, data, showSMSCheckbox) => {
+  if (type === 'phone') {
+    let initialFormValues = {
+      countryCode: '1',
+      extension: '',
+      inputPhoneNumber: '',
+      isTextable: false,
+      isTextPermitted: false,
+      'view:showSMSCheckbox': showSMSCheckbox,
+    };
+
+    if (data) {
+      const { extension, areaCode, phoneNumber, isTextPermitted } = data;
+
+      initialFormValues = {
+        ...data,
+        extension: extension || '',
+        inputPhoneNumber: `${areaCode}${phoneNumber}`,
+        isTextPermitted: isTextPermitted || false,
+        'view:showSMSCheckbox': showSMSCheckbox,
+      };
+    }
+
+    return initialFormValues;
+  }
+
+  return null;
+};
